% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseConnection.R
\docType{data}
\name{Connection}
\alias{Connection}
\title{Global: Build a database connection to the AK_ShallowLakes database.}
\format{
An object of class \code{Microsoft SQL Server} of length 1.
}
\usage{
Connection
}
\description{
Global: Build a database connection to the AK_ShallowLakes database.
}
\keyword{datasets}
