% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDatabaseConnection.R
\name{GetDatabaseConnection}
\alias{GetDatabaseConnection}
\title{GetDatabaseConnection()
This function builds a database connection to the AK_ShallowLakes database.}
\usage{
GetDatabaseConnection()
}
\value{
An odbc database connection object.
}
\description{
GetDatabaseConnection()
This function builds a database connection to the AK_ShallowLakes database.
}
\examples{
Connection = GetDatabaseConnection()    MyDataFrame = odbc::dbGetQuery(Connection,'SELECT Top 3 * FROM tblPonds')
}
