% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLakeMap.R
\name{GetLakeCoordinates}
\alias{GetLakeCoordinates}
\title{GetLakeCoordinates}
\usage{
GetLakeCoordinates(Lake)
}
\arguments{
\item{Lake}{Lake for which spatial coordinates should be returned}
}
\value{
A plot
}
\description{
This returns a plot of spatial coordinates for a shallow lakes monitoring lake in geographic coordinate system.
}
\examples{
# Get a plot of Lake
GetLakeCoordinates('YUCH-004')
}
