% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLakeMap.R
\name{GetLakeMap}
\alias{GetLakeMap}
\title{GetLakeMap}
\usage{
GetLakeMap(Lake)
}
\arguments{
\item{Lake}{Lake for which spatial coordinates should be returned}
}
\value{
A Leaflet map for Lake.
}
\description{
Function to retrieve the spatial coordinates of Lake from the AK_ShallowLakes:tblMonuments table and display them in a Leaflet map.
}
\examples{
# Get a Leaflet map showing Lake YUCH-004
GetLakeMap('YUCH-004')
}
