% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLakesDataset.R
\name{GetLakesDataset}
\alias{GetLakesDataset}
\title{GetLakesDataset}
\usage{
GetLakesDataset()
}
\value{
A data frame of Lakes data.
}
\description{
Queries the [tblPonds] database table and returns the records as a data frame.
}
\examples{
# Example of use:
data = GetLakesDataset()
head(data)
}
