% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWaterProfilesDataset.R
\name{GetWaterProfilesDataset}
\alias{GetWaterProfilesDataset}
\title{GetWaterProfilesDataset}
\usage{
GetWaterProfilesDataset()
}
\value{
A data frame of water profile data.
}
\description{
Queries the [AK_ShallowLakes].[dbo].[tblWaterProfiles] database table and returns the records as a data frame.
}
\examples{
# Example of use:
data = GetWaterProfilesDataset()
head(data)
}
