% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWaterProfilesDataSummary.R
\name{GetWaterProfilesDataSummary}
\alias{GetWaterProfilesDataSummary}
\title{GetWaterProfilesDataSummary}
\usage{
GetWaterProfilesDataSummary()
}
\value{
A data frame of water profile statistics for temperature, pH, and dissolved oxygen.
}
\description{
The GetWaterProfilesDataSummary function returns a summary of water profile parameters over sample replicates (usually A,B,C,etc.) and grouped by Lake and sampling date.
Statistics include mean, min, max, SD and n) for each lake and sampling date.
The statistics returned by this function are calculated by the AK_ShallowLakes database in the view (query) [Summary_WaterProfileStatisticsByLake].
Results are returned as a data frame.
}
\examples{
# Example of use:
data = GetWaterProfilesDataSummary
head(data)
}
