% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterProfiles_GetDataSummary.R
\name{WaterProfiles_GetDataSummary}
\alias{WaterProfiles_GetDataSummary}
\title{WaterProfiles_GetDataSummary}
\usage{
WaterProfiles_GetDataSummary()
}
\value{
A data frame of water profile statistics for temperature, pH, and dissolved oxygen.
}
\description{
Queries the [AK_ShallowLakes].[dbo].[Summary_WaterProfileStatisticsByLake] view and returns the records as a data frame. Data is summarized over SAMPLENUMBER replicates by lake and date.
}
\examples{
# Example of use:
WaterProfilesDataSummary = WaterProfiles_GetDataSummary
head(WaterProfilesDataSummary)
}
